requirements.txt
python-telegram-bot==12.7
googletrans==3.1.0a0